modularized further
