+++ src/seriespl.cpp
@@ -24,8 +24,7 @@
 #include <iterator>
 #include <vector>
 #include <sstream>
-#include <unistd.h>
-#include <getopt.h>
+#include <unistd.h>	// sleep()
 #include <map>
 #include <utility>
 #include <cstdio>
@@ -34,42 +33,40 @@
 #include "http.hpp"
 #include "config.hpp"
 
-Seriespl::Seriespl()
+Seriespl::Seriespl(int argc, char const *argv[])
+	// Set default list of active streaming providers, SSL only
+:	Providers({ Streamcloud, Vivo, Shared, YouTube, OpenLoad }),
+	providermap
+	({
+		{Streamcloud, providerpair("Streamcloud", "streamcloud.eu")},
+		{Vivo, providerpair("Vivo", "vivo.sx")},
+		{Shared, providerpair("Shared", "shared.sx")},
+		{YouTube, providerpair("YouTube", "www.youtube.com")},
+		{OpenLoad, providerpair("OpenLoad", "openload.co")},
+		{PowerWatch, providerpair("PowerWatch", "powerwatch.pw")},
+		{CloudTime, providerpair("CloudTime", "www.cloudtime.to")},
+		{AuroraVid, providerpair("AuroraVid", "auroravid.to")},
+		{Vidto, providerpair("Vidto", "vidto.me")}
+	})
 {
-	Providers =
-	{ // Set default list of active streaming providers, SSL only
-		Streamcloud,
-		Vivo,
-		Shared,
-		YouTube,
-		OpenLoad
-	};
-}
-
-Seriespl::~Seriespl() {}
-
-int Seriespl::start()
-{
-	if (directoryurl.find("bs.to") != std::string::npos)
-	{
-		service = BurningSeries;
-	}
-	else
-	{
-		std::cerr << "Error: Could not determine which website you specified, " <<
-			"defaulting to Burning-Series." << std::endl;
-		sleep(2);
-		service = BurningSeries;
-	}
-
 	// read config and set streaming providers, if specified
-	if (readconfig(config))
+	if (Config::read(config))
 	{
 		if (config["streamproviders"] != "")
 			populate_providers(config["streamproviders"]);
 		if (config["youtube-dl"] != "")
 			yt_dl_path = config["youtube-dl"];
 	}
+	handle_args(argc, argv);
+
+	set_service();
+}
+
+Seriespl::~Seriespl() {}
+
+int Seriespl::run()
+{
+	std::string content;
 
 	if (service == BurningSeries)
 	{
@@ -103,11 +100,11 @@ int Seriespl::start()
 				{
 					directoryurl = directoryurl + "/" + std::to_string(season);
 				}
-				content = getpage(directoryurl);
+				content = GetHTTP::getpage(directoryurl);
 			}
 			else
 			{ // If no season range was selected, use supplied URL
-				content = getpage(directoryurl);
+				content = GetHTTP::getpage(directoryurl);
 			}
 
 			if (current_episode != 0)
@@ -196,192 +193,19 @@ int Seriespl::start()
 	return 0;
 }
 
-int Seriespl::handle_args(int argc, char const *argv[])
+void Seriespl::set_service()
 {
-	int opt;
-	std::string usage = std::string("usage: ") + argv[0] +
-		" [-h] [-i]|[-p list] [-e episodes] [-s seasons] [-f format] [-y] URL";
-	
-	while ((opt = getopt(argc, (char **)argv, "hp:ie:s:f:yV")) != -1)
+	if (directoryurl.find("bs.to") != std::string::npos)
 	{
-		std::string episodes, seasons;
-		size_t pos;
-		switch (opt)
-		{
-			case 'h':	// Help
-				std::cout << usage << std::endl << std::endl;
-				std::cout <<
-					"  -h                   Show this help" << std::endl;
-				std::cout <<
-					"  -p stream providers  Comma delimited list. Available:" << std::endl;
-				std::cout <<
-					"                       Streamcloud,Vivo,Shared,YouTube,OpenLoad,PowerWatch,CloudTime,AuroraVid,Vidto" << std::endl;
-				std::cout <<
-					"  -i                   Use stream providers without SSL support too" << std::endl;
-				std::cout <<
-					"  -e                   Episode range, e.g. 2-5 or 7 or 9-, use c for current" << std::endl;
-				std::cout <<
-					"  -s                   Season range, e.g. 1-2 or 4" << std::endl;
-				std::cout <<
-					"  -f                   Playlist format. Available: raw, m3u, pls" << std::endl;
-				std::cout <<
-					"  -y                   Use youtube-dl to print the direct URL of the video file" << std::endl;
-				std::cout <<
-					"  -V                   Output version and copyright information and exit" << std::endl;
-				return 0;
-				break;
-			case 'p':	// Provider
-				populate_providers(std::string(optarg));
-				break;
-			case 'i':	// Insecure
-				Providers =
-				{
-					Streamcloud,
-					Vivo,
-					Shared,
-					YouTube,
-					OpenLoad,
-					PowerWatch,
-					CloudTime,
-					AuroraVid,
-					Vidto
-				};
-				break;
-			case 'e':	// Episodes
-				episodes = optarg;
-				pos = episodes.find('-');
-				if (pos != std::string::npos)
-				{
-					try
-					{
-						if (episodes.substr(0, pos) == "c")
-						{
-							current_episode = 1;
-						}
-						else
-						{
-							startEpisode = std::stoi( episodes.substr(0, pos) );
-						}
-						if (episodes.length() > pos + 1)
-						{ // If episodes = 5-, output all episodes, beginning with 5
-							if (episodes.substr(pos + 1) == "c")
-							{
-								if (current_episode == 1)
-								{
-									current_episode = 3;
-								}
-								else
-								{
-									current_episode = 2;
-								}
-							}
-							else
-							{
-								endEpisode = std::stoi( episodes.substr(pos + 1) );
-							}
-						}
-					}
-					catch (std::exception &e)
-					{ // There is a '-' but no numbers around it
-						std::cerr << "Error: Can not decipher episode range, " <<
-							"defaulting to all." << std::endl;
-						sleep(2);
-					}
-				}
-				else
-				{
-					try
-					{ 
-						if (episodes == "c")
-						{
-							current_episode = 3;
-						}
-						else
-						{ // Is episodes a single number?
-							startEpisode = std::stoi(episodes);
-							endEpisode = std::stoi(episodes);
-						}
-					}
-					catch (std::exception &e)
-					{
-						std::cerr << "Error: Can not decipher episode range, " <<
-							"defaulting to all." << std::endl;
-						sleep(2);
-					}
-				}
-				break;
-			case 's':	// Seasons
-				seasons = optarg;
-				pos = seasons.find('-');
-				if (pos != std::string::npos)
-				{
-					try
-					{
-						startSeason = std::stoi( seasons.substr(0, pos) );
-						endSeason = std::stoi( seasons.substr(pos + 1) );
-					}
-					catch (std::exception &e)
-					{ // There is a '-' but no numbers around it
-						std::cerr << "Error: Can not decipher season range, " <<
-							"defaulting to selected." << std::endl;
-						startSeason = -1;
-						endSeason = -1;
-						sleep(2);
-					}
-				}
-				else
-				{
-					try
-					{ // Is seasons a single number?
-						startSeason = std::stoi(seasons);
-						endSeason = std::stoi(seasons);
-					}
-					catch (std::exception &e)
-					{
-						std::cerr << "Error: Can not decipher season range, " <<
-							"defaulting to selected." << std::endl;
-						sleep(2);
-					}
-				}
-				break;
-			case 'f':	// Format
-				if (strncmp(optarg, "raw", 3) == 0)
-					playlist = PL_RAW;
-				else if (strncmp(optarg, "m3u", 3) == 0)
-					playlist = PL_M3U;
-				else if (strncmp(optarg, "pls", 3) == 0)
-					playlist = PL_PLS;
-				else
-					std::cerr << "Playlist format not recognized, defaulting to raw." << std::endl;
-				break;
-			case 'y':	// youtube-dl
-				direct_url = true;
-				break;
-			case 'V':	// Version
-				std::cout << "seriespl " << version << "\n"
-						  << "Copyright © 2016 tastytea <tastytea@tastytea.de>.\n"
-						  << "License GPLv2: GNU GPL version 2 <http://www.gnu.org/licenses/gpl-2.0.html>.\n"
-						  << "This is free software: you are free to change and redistribute it.\n"
-						  << "There is NO WARRANTY, to the extent permitted by law." << std::endl;
-				return 0;
-				break;
-			default:
-				std::cerr << usage << std::endl;
-				return 1;
-				break;
-		}
-	}
-	if (optind >= argc)
-	{
-		std::cerr << usage << std::endl;
-		return 1;
+		service = BurningSeries;
 	}
 	else
-	{ // Set URL of tv series
-		directoryurl = argv[optind];
+	{
+		std::cerr << "Error: Could not determine which website you specified, " <<
+			"defaulting to Burning-Series." << std::endl;
+		sleep(2);
+		service = BurningSeries;
 	}
-
-	return 0;
 }
 
 void Seriespl::populate_providers(const std::string &providerlist)
@@ -415,7 +239,7 @@ void Seriespl::populate_providers(const std::string &providerlist)
 std::string Seriespl::getlink(const std::string &url, const StreamProviders &provider,
 							  std::string &title)
 { // Takes URL of episode-page and streaming provider, returns URL of stream-page or "" on error
-	std::string content = getpage(url);
+	std::string content = GetHTTP::getpage(url);
 	std::string streamurl = "";
 
 	if (service == BurningSeries)
@@ -463,58 +287,6 @@ std::string Seriespl::getlink(const std::string &url, const StreamProviders &pro
 {
 	std::string title;
 	return getlink(url, provider, title);
-}
-
-void Seriespl::print_playlist(const PlaylistFormat &playlist, const std::string &url,
-							  const std::string &title)
-{
-	static unsigned short counter = 1;
-	std::string newtitle = title;
-	size_t pos = 0;
-
-	switch (playlist)
-	{
-		case PL_RAW:
-			std::cout << url << std::endl;
-			break;
-		case PL_M3U:
-			if (counter == 1) // Write header
-			{
-				std::cout << "#EXTM3U" << std::endl;
-			}
-			// Replacing comma with U+201A, SINGLE LOW-9 QUOTATION MARK
-			// Because VLC uses commas in titles as separator for metadata
-			while ((pos = title.find(',', pos + 1)) != std::string::npos)
-			{
-				newtitle.replace(pos, 1, "‚");	// The ‚ is not a comma
-			}
-			std::cout << "#EXTINF:-1," << newtitle << std::endl;
-			std::cout << url << std::endl;
-			break;
-		case PL_PLS:
-			if (counter == 1) // Write header
-			{
-				std::cout << "[playlist]" << std::endl;
-				std::cout << "Version=2" << std::endl;
-			}
-			if (url == "NUMBER_OF_EPISODES")
-			{
-				std::cout << "NumberOfEntries=" << counter - 1 << std::endl;
-			}
-			else
-			{
-				std::cout << "File" << counter << "=" << url << std::endl;
-				std::cout << "Title" << counter << "=" << title << std::endl;
-				std::cout << "Length" << counter << "=-1" << std::endl;
-			}
-			break;
-	}
-	++counter;
-}
-
-void Seriespl::print_playlist(const PlaylistFormat &playlist, const std::string &url)
-{
-	return print_playlist(playlist, url, "");
 }
 
 std::string Seriespl::get_direct_url(std::string &providerurl)
